+++ tests/ota/test_ota_providers.py
@@ -10,7 +10,6 @@ import attrs
 import pytest
 
 from tests.conftest import make_node_desc
-from tests.ota.test_ota_metadata import image_with_metadata  # noqa: F401
 import zigpy.device
 from zigpy.ota import OtaImageWithMetadata, providers
 import zigpy.types as t
