+++ tests/test_appdb_migration.py
@@ -8,7 +8,7 @@ from aiosqlite.context import contextmanager
 import pytest
 
 from tests.async_mock import AsyncMock, MagicMock, patch
-from tests.conftest import app, make_node_desc  # noqa: F401
+from tests.conftest import make_node_desc
 from tests.test_appdb import make_app_with_db
 import zigpy.appdb
 import zigpy.appdb_schemas
